chore: restructure service files
